add cavity, magres, latticeplane, crystalshape to plugins
