+++ batoms/crystal_shape/ui_list_crystal_shape.py
@@ -5,40 +5,40 @@ import bpy
 from bpy.types import Menu, Panel, UIList
 
 
-class BATOMS_MT_lattice_plane_context_menu(Menu):
-    bl_label = "Lattice Plane Specials"
-    bl_idname = "BATOMS_MT_lattice_plane_context_menu"
+class BATOMS_MT_crystal_shape_context_menu(Menu):
+    bl_label = "Crystal Shape Specials"
+    bl_idname = "BATOMS_MT_crystal_shape_context_menu"
 
     def draw(self, _context):
         layout = self.layout
-        op = layout.operator("plane.lattice_plane_add",
-                             icon='ADD', text="Add Lattice Plane")
+        op = layout.operator("plane.crystal_shape_add",
+                             icon='ADD', text="Add Crystal Shape")
         layout.separator()
-        op = layout.operator("plane.lattice_plane_remove",
-                             icon='X', text="Delete All Lattice Plane")
+        op = layout.operator("plane.crystal_shape_remove",
+                             icon='X', text="Delete All Crystal Shape")
         op.all = True
 
 
-class BATOMS_UL_lattice_planes(UIList):
+class BATOMS_UL_crystal_shape(UIList):
     def draw_item(self, _context, layout, _data, item, icon, active_data, _active_propname, index):
-        lattice_plane = item
+        crystal_shape = item
         custom_icon = 'OBJECT_DATAMODE'
         if self.layout_type in {'DEFAULT', 'COMPACT'}:
             split = layout.split(factor=0.66, align=False)
-            split.prop(lattice_plane, "name", text="",
+            split.prop(crystal_shape, "name", text="",
                        emboss=False, icon=custom_icon)
             row = split.row(align=True)
             row.emboss = 'NONE_OR_STATUS'
-            row.prop(lattice_plane, "distance", text="")
+            row.prop(crystal_shape, "distance", text="")
         elif self.layout_type == 'GRID':
             layout.alignment = 'CENTER'
             layout.label(text="", icon=custom_icon)
 
 
-class BATOMS_PT_lattice_planes(Panel):
-    bl_label = "Lattice Plane"
+class BATOMS_PT_crystal_shape(Panel):
+    bl_label = "Crystal Shape"
     bl_category = "Surface"
-    bl_idname = "BATOMS_PT_Lattice_Planes"
+    bl_idname = "BATOMS_PT_crystal_shape"
     bl_space_type = "VIEW_3D"
     bl_region_type = "UI"
     # bl_options = {'DEFAULT_CLOSED'}
@@ -58,8 +58,8 @@ class BATOMS_PT_lattice_planes(Panel):
 
         ob = context.object
         ba = bpy.data.collections[ob.batoms.label].batoms
-        if len(ba.blatticeplane) > 0:
-            kb = ba.blatticeplane[ba.latticeplane_index]
+        if len(ba.bcrystalshape) > 0:
+            kb = ba.bcrystalshape[ba.crystalshape_index]
         else:
             kb = None
 
@@ -69,17 +69,17 @@ class BATOMS_PT_lattice_planes(Panel):
         if kb:
             rows = 5
 
-        row.template_list("BATOMS_UL_lattice_planes", "", ba,
-                          "blatticeplane", ba, "latticeplane_index", rows=rows)
+        row.template_list("BATOMS_UL_crystal_shape", "", ba,
+                          "bcrystalshape", ba, "crystalshape_index", rows=rows)
 
         col = row.column(align=True)
-        op = col.operator("plane.lattice_plane_add", icon='ADD', text="")
-        op = col.operator("plane.lattice_plane_remove", icon='REMOVE', text="")
+        op = col.operator("plane.crystal_shape_add", icon='ADD', text="")
+        op = col.operator("plane.crystal_shape_remove", icon='REMOVE', text="")
         if kb is not None:
             op.name = kb.name
         col.separator()
 
-        col.menu("BATOMS_MT_lattice_plane_context_menu",
+        col.menu("BATOMS_MT_crystal_shape_context_menu",
                  icon='DOWNARROW_HLT', text="")
 
         if kb:
@@ -102,9 +102,10 @@ class BATOMS_PT_lattice_planes(Panel):
             col = layout.column()
             sub = col.column(align=True)
             sub.prop(kb, "distance", text="Distance")
-            sub.prop(kb, "scale", text="Scale")
+            # sub.prop(kb, "scale", text="Scale")
+            col.prop(kb, "symmetry",  text="Symmetry")
             col.prop(kb, "show_edge",  text="Show edge")
             col.prop(kb, "material_style", text="material_style")
             col.prop(kb, "color",  text="color")
-            op = layout.operator("plane.lattice_plane_draw",
+            op = layout.operator("plane.crystal_shape_draw",
                                  icon='GREASEPENCIL', text="Draw")
