+++ batoms/gui/__init__.py
@@ -15,12 +15,8 @@ from . import (
     ui_list_species,
     ui_list_bond,
     ui_list_polyhedra,
-    ui_list_lattice_plane,
-    ui_list_crystal_shape,
     ui_list_isosurface,
-    ui_list_cavity,
     ui_list_ms,
-    ui_list_magres,
     view3d_mt_batoms_add,
     view3d_mt_object_context_menu,
     view3d_mt_edit_mesh_context_menu,
@@ -85,24 +81,12 @@ classes = [
     ui_list_polyhedra.BATOMS_MT_polyhedra_context_menu,
     ui_list_polyhedra.BATOMS_UL_polyhedra,
     ui_list_polyhedra.BATOMS_PT_polyhedra,
-    ui_list_lattice_plane.BATOMS_MT_lattice_plane_context_menu,
-    ui_list_lattice_plane.BATOMS_UL_lattice_planes,
-    ui_list_lattice_plane.BATOMS_PT_lattice_planes,
-    ui_list_crystal_shape.BATOMS_MT_crystal_shape_context_menu,
-    ui_list_crystal_shape.BATOMS_UL_crystal_shapes,
-    ui_list_crystal_shape.BATOMS_PT_crystal_shapes,
     ui_list_isosurface.BATOMS_MT_isosurface_context_menu,
     ui_list_isosurface.BATOMS_UL_isosurface,
     ui_list_isosurface.BATOMS_PT_isosurface,
-    ui_list_cavity.BATOMS_MT_cavity_context_menu,
-    ui_list_cavity.BATOMS_UL_cavity,
-    ui_list_cavity.BATOMS_PT_cavity,
     ui_list_ms.BATOMS_MT_ms_context_menu,
     ui_list_ms.BATOMS_UL_ms,
     ui_list_ms.BATOMS_PT_ms,
-    ui_list_magres.BATOMS_MT_magres_context_menu,
-    ui_list_magres.BATOMS_UL_magres,
-    ui_list_magres.BATOMS_PT_magres,
     BatomsCollection,
 ]
