+++ batoms/lattice_plane/ui_list_lattice_plane.py
@@ -5,40 +5,40 @@ import bpy
 from bpy.types import Menu, Panel, UIList
 
 
-class BATOMS_MT_crystal_shape_context_menu(Menu):
-    bl_label = "Crystal Shape Specials"
-    bl_idname = "BATOMS_MT_crystal_shape_context_menu"
+class BATOMS_MT_lattice_plane_context_menu(Menu):
+    bl_label = "Lattice Plane Specials"
+    bl_idname = "BATOMS_MT_lattice_plane_context_menu"
 
     def draw(self, _context):
         layout = self.layout
-        op = layout.operator("plane.crystal_shape_add",
-                             icon='ADD', text="Add Crystal Shape")
+        op = layout.operator("plane.lattice_plane_add",
+                             icon='ADD', text="Add Lattice Plane")
         layout.separator()
-        op = layout.operator("plane.crystal_shape_remove",
-                             icon='X', text="Delete All Crystal Shape")
+        op = layout.operator("plane.lattice_plane_remove",
+                             icon='X', text="Delete All Lattice Plane")
         op.all = True
 
 
-class BATOMS_UL_crystal_shapes(UIList):
+class BATOMS_UL_lattice_plane(UIList):
     def draw_item(self, _context, layout, _data, item, icon, active_data, _active_propname, index):
-        crystal_shape = item
+        lattice_plane = item
         custom_icon = 'OBJECT_DATAMODE'
         if self.layout_type in {'DEFAULT', 'COMPACT'}:
             split = layout.split(factor=0.66, align=False)
-            split.prop(crystal_shape, "name", text="",
+            split.prop(lattice_plane, "name", text="",
                        emboss=False, icon=custom_icon)
             row = split.row(align=True)
             row.emboss = 'NONE_OR_STATUS'
-            row.prop(crystal_shape, "distance", text="")
+            row.prop(lattice_plane, "distance", text="")
         elif self.layout_type == 'GRID':
             layout.alignment = 'CENTER'
             layout.label(text="", icon=custom_icon)
 
 
-class BATOMS_PT_crystal_shapes(Panel):
-    bl_label = "Crystal Shape"
+class BATOMS_PT_lattice_plane(Panel):
+    bl_label = "Lattice Plane"
     bl_category = "Surface"
-    bl_idname = "BATOMS_PT_crystal_shapes"
+    bl_idname = "BATOMS_PT_Lattice_Plane"
     bl_space_type = "VIEW_3D"
     bl_region_type = "UI"
     # bl_options = {'DEFAULT_CLOSED'}
@@ -58,8 +58,8 @@ class BATOMS_PT_crystal_shapes(Panel):
 
         ob = context.object
         ba = bpy.data.collections[ob.batoms.label].batoms
-        if len(ba.bcrystalshape) > 0:
-            kb = ba.bcrystalshape[ba.crystalshape_index]
+        if len(ba.blatticeplane) > 0:
+            kb = ba.blatticeplane[ba.latticeplane_index]
         else:
             kb = None
 
@@ -69,17 +69,17 @@ class BATOMS_PT_crystal_shapes(Panel):
         if kb:
             rows = 5
 
-        row.template_list("BATOMS_UL_crystal_shapes", "", ba,
-                          "bcrystalshape", ba, "crystalshape_index", rows=rows)
+        row.template_list("BATOMS_UL_lattice_plane", "", ba,
+                          "blatticeplane", ba, "latticeplane_index", rows=rows)
 
         col = row.column(align=True)
-        op = col.operator("plane.crystal_shape_add", icon='ADD', text="")
-        op = col.operator("plane.crystal_shape_remove", icon='REMOVE', text="")
+        op = col.operator("plane.lattice_plane_add", icon='ADD', text="")
+        op = col.operator("plane.lattice_plane_remove", icon='REMOVE', text="")
         if kb is not None:
             op.name = kb.name
         col.separator()
 
-        col.menu("BATOMS_MT_crystal_shape_context_menu",
+        col.menu("BATOMS_MT_lattice_plane_context_menu",
                  icon='DOWNARROW_HLT', text="")
 
         if kb:
@@ -102,10 +102,9 @@ class BATOMS_PT_crystal_shapes(Panel):
             col = layout.column()
             sub = col.column(align=True)
             sub.prop(kb, "distance", text="Distance")
-            # sub.prop(kb, "scale", text="Scale")
-            col.prop(kb, "symmetry",  text="Symmetry")
+            sub.prop(kb, "scale", text="Scale")
             col.prop(kb, "show_edge",  text="Show edge")
             col.prop(kb, "material_style", text="material_style")
             col.prop(kb, "color",  text="color")
-            op = layout.operator("plane.crystal_shape_draw",
+            op = layout.operator("plane.lattice_plane_draw",
                                  icon='GREASEPENCIL', text="Draw")
