+++ batoms/magres/ui_list_magres.py
