+++ batoms/ops/__init__.py
@@ -10,12 +10,10 @@ from . import (
     ops_batoms,
     ops_species,
     ops_surface,
-    ops_cavity,
     measure,
     manual_mapping,
     ops_bond,
     ops_polyhedra,
-    ops_plane,
     ops_render,
     )
 
@@ -63,6 +61,7 @@ classes = [
     ops_batoms.BatomsSeparate,
     ops_batoms.deleteBatoms,
     ops_batoms.deleteSelectedBatoms,
+    ops_batoms.BatomsDeleteSelectedAtoms,
     ops_io.IMPORT_OT_batoms,
     ops_io.EXPORT_OT_batoms,
     ops_species.SpeciesAdd,
@@ -80,28 +79,14 @@ classes = [
     ops_polyhedra.PolyhedraRemove,
     ops_polyhedra.PolyhedraDraw,
     ops_polyhedra.PolyhedraModify,
-    ops_plane.LatticePlaneAdd,
-    ops_plane.LatticePlaneRemove,
-    ops_plane.LatticePlaneDraw,
-    ops_plane.LatticePlaneModify,
-    ops_plane.CrystalShapeAdd,
-    ops_plane.CrystalShapeRemove,
-    ops_plane.CrystalShapeDraw,
-    ops_plane.CrystalShapeModify,
     ops_surface.MSAdd,
     ops_surface.MSRemove,
     ops_surface.MSDraw,
     ops_surface.MSModify,
-    ops_surface.MagresAdd,
-    ops_surface.MagresRemove,
-    ops_surface.MagresDraw,
     ops_surface.IsosurfaceAdd,
     ops_surface.IsosurfaceRemove,
     ops_surface.IsosurfaceDraw,
     ops_surface.IsosurfaceModify,
-    ops_cavity.CavityAdd,
-    ops_cavity.CavityRemove,
-    ops_cavity.CavityDraw,
     ops_render.RenderAdd,
     measure.MeasureButton,
 ]
